requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.1

langchain>=0.1.17
langchain-openai>=0.1.6
langchain-community>=0.0.36
langchain-text-splitters>=0.0.1
pypdf>=4.2.0

openai>=1.23.0

faiss-cpu>=1.8.0
tiktoken>=0.6.0

pydantic>=2.6.4
loguru>=0.7.2
